--- src/main/java/DukeException.java
@@ -1,5 +1,0 @@
-public class DukeException extends IllegalArgumentException {
-    public DukeException(String errorMessage) {
-        super(errorMessage);
-    }
-}
