+++ src/main/java/duke/main/DukeException.java
@@ -1,0 +1,7 @@
+package duke.main;
+
+public class DukeException extends IllegalArgumentException {
+    public DukeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
