+++ src/main/java/duke/main/Storage.java
@@ -1,3 +1,7 @@
+package duke.main;
+
+import duke.task.*;
+
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
