--- text-ui-test/ACTUAL.TXT
@@ -1,54 +1,0 @@
-Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
-
-	Nice to see you again.
-	 You have no tasks in the list.
-
-	I don't understand that command (yet...)
-
-	You haven't added any tasks yet
-
-	☹ OOPS!!! Your todo needs a description.
-
-	 Got it. I've added this task:
-	 	 [T][ ] do one thing
-	 You have 1 task in the list.
-
-	☹ OOPS!!! You need to specify a time.
-
-	☹ OOPS!!! Please specify the time in the yyyy-mm-dd format.
-
-	 Got it. I've added this task:
-	 	 [D][ ] do 2 things (by: Oct 12 2020)
-	 You have 2 tasks in the list.
-
-	☹ OOPS!!! You need to specify a time.
-
-	☹ OOPS!!! Please specify the time in the yyyy-mm-dd format.
-
-	 Got it. I've added this task:
-	 	 [E][ ] a birthday (at: Oct 12 2020)
-	 You have 3 tasks in the list.
-
-	Sorry, I can't seem to find that task
-
-	 Got it. I've removed this task:
-	 	 [E][ ] a birthday (at: Oct 12 2020)
-	 You have 2 tasks in the list.
-
-	I'm Sorry, WHAT?!?!
-
-	 Nice! I've marked this task as done:
-		 [D][X] do 2 things (by: Oct 12 2020)
-
-	 1. [T][ ] do one thing
-	 2. [D][X] do 2 things (by: Oct 12 2020)
-
-	Clearing tasks...
-
-	You can now start anew...
-
